Move query related functions to its own file
